+++ tests/test_features_beats.py
@@ -8,23 +8,35 @@ This file contains tests for the madmom.ml.hmm module.
 from __future__ import absolute_import, division, print_function
 
 import unittest
-from . import ACTIVATIONS_PATH, MODELS_PATH
+from . import AUDIO_PATH, ACTIVATIONS_PATH, MODELS_PATH
 from madmom.ml.hmm import HiddenMarkovModel
+from madmom.features import Activations
 from madmom.features.beats import *
 from madmom.features.beats_hmm import *
 
-act_file = np.load("%s/sample.beats_blstm_2013.npz" % ACTIVATIONS_PATH)
-fps = act_file['fps']
-act = act_file['activations']
+sample_file = "%s/sample.wav" % AUDIO_PATH
+sample_act = Activations("%s/sample.beats_blstm_2013.npz" % ACTIVATIONS_PATH)
+sample_mb_features = Activations("%s/sample.multiband_spectral_flux.npz" %
+                                 ACTIVATIONS_PATH)
+
+
+class TestRNNBeatProcessorClass(unittest.TestCase):
+
+    def setUp(self):
+        self.processor = RNNBeatProcessor()
+
+    def test_process(self):
+        beat_act = self.processor(sample_file)
+        self.assertTrue(np.allclose(beat_act, sample_act))
 
 
 class TestBeatTrackingProcessorClass(unittest.TestCase):
 
     def setUp(self):
-        self.processor = BeatTrackingProcessor(fps=fps)
+        self.processor = BeatTrackingProcessor(fps=sample_act.fps)
 
     def test_process(self):
-        beats = self.processor(act)
+        beats = self.processor(sample_act)
         self.assertTrue(np.allclose(beats, [0.11, 0.45, 0.79, 1.13, 1.47,
                                             1.81, 2.15, 2.49]))
 
@@ -32,10 +44,10 @@ class TestBeatTrackingProcessorClass(unittest.TestCase):
 class TestBeatDetectionProcessorClass(unittest.TestCase):
 
     def setUp(self):
-        self.processor = BeatDetectionProcessor(fps=fps)
+        self.processor = BeatDetectionProcessor(fps=sample_act.fps)
 
     def test_process(self):
-        beats = self.processor(act)
+        beats = self.processor(sample_act)
         self.assertTrue(np.allclose(beats, [0.11, 0.45, 0.79, 1.13, 1.47,
                                             1.81, 2.15, 2.49]))
 
@@ -43,17 +55,17 @@ class TestBeatDetectionProcessorClass(unittest.TestCase):
 class TestCRFBeatDetectionProcessorClass(unittest.TestCase):
 
     def setUp(self):
-        self.processor = CRFBeatDetectionProcessor(fps=fps)
+        self.processor = CRFBeatDetectionProcessor(fps=sample_act.fps)
 
     def test_process(self):
-        beats = self.processor(act)
+        beats = self.processor(sample_act)
         self.assertTrue(np.allclose(beats, [0.09, 0.79, 1.49]))
 
 
 class TestDBNBeatTrackingProcessorClass(unittest.TestCase):
 
     def setUp(self):
-        self.processor = DBNBeatTrackingProcessor(fps=fps)
+        self.processor = DBNBeatTrackingProcessor(fps=sample_act.fps)
 
     def test_types(self):
         self.assertIsInstance(self.processor.correct, bool)
@@ -65,7 +77,7 @@ class TestDBNBeatTrackingProcessorClass(unittest.TestCase):
 
     def test_values(self):
         self.assertTrue(self.processor.correct)
-        path, prob = self.processor.hmm.viterbi(act)
+        path, prob = self.processor.hmm.viterbi(sample_act)
         self.assertTrue(np.allclose(path[:15], [2030, 2031, 2032, 2033, 2034,
                                                 2035, 2036, 1968, 1969, 1970,
                                                 1971, 1972, 1973, 1974, 1975]))
@@ -77,7 +89,7 @@ class TestDBNBeatTrackingProcessorClass(unittest.TestCase):
                                      0.98550725, 0, 0.01449275, 0.02898551]))
 
     def test_process(self):
-        beats = self.processor(act)
+        beats = self.processor(sample_act)
         self.assertTrue(np.allclose(beats, [0.09, 0.8, 1.48, 2.15]))
 
 
@@ -86,11 +98,8 @@ class TestPatternTrackingProcessorClass(unittest.TestCase):
     def setUp(self):
         import glob
         pattern_files = sorted(glob.glob("%s/patterns/2013/*" % MODELS_PATH))
-        features = np.load("%s/sample.multiband_spectral_flux.npz"
-                           % ACTIVATIONS_PATH)
-        self.act = features['activations']
         self.processor = PatternTrackingProcessor(pattern_files,
-                                                  fps=features['fps'])
+                                                  fps=sample_mb_features.fps)
 
     def test_types(self):
         self.assertIsInstance(self.processor.downbeats, bool)
@@ -105,20 +114,21 @@ class TestPatternTrackingProcessorClass(unittest.TestCase):
         self.assertTrue(self.processor.fps == 50)
         self.assertTrue(self.processor.downbeats is False)
         self.assertTrue(np.allclose(self.processor.num_beats, [3, 4]))
-        path, prob = self.processor.hmm.viterbi(self.act)
+        path, prob = self.processor.hmm.viterbi(sample_mb_features)
         self.assertTrue(np.allclose(path[:12], [5573, 5574, 5575, 5576, 6757,
                                                 6758, 6759, 6760, 6761, 6762,
                                                 6763, 6764]))
         self.assertTrue(np.allclose(prob, -468.8014))
         patterns = self.processor.st.state_patterns[path]
-        self.assertTrue(np.allclose(patterns, np.ones(len(self.act))))
+        self.assertTrue(np.allclose(patterns,
+                                    np.ones(len(sample_mb_features))))
         positions = self.processor.st.state_positions[path]
         self.assertTrue(np.allclose(positions[:6], [1.76470588, 1.82352944,
                                                     1.88235296, 1.94117648,
                                                     2, 2.0588236]))
 
     def test_process(self):
-        beats = self.processor(self.act)
+        beats = self.processor(sample_mb_features)
         self.assertTrue(np.allclose(beats, [[0.08, 3], [0.42, 4], [0.76, 1],
                                             [1.1, 2], [1.44, 3], [1.78, 4],
                                             [2.12, 1], [2.46, 2], [2.8, 3]]))