+++ CHANGES.rst
@@ -8,6 +8,17 @@ API relevant changes:
 
 * Refactored the `ml.rnn` to `ml.nn` and converted the models to pickles (#110)
 
+Other changes:
+
+* added classes for onset/note/beat detection with RNNs to `features.*` (#118)
+
+Version 0.13.1 (release date: 2016-03-14)
+-----------------------------------------
+
+This is a bugfix release.
+
+* Fix beat evaluation argument parsing (#116)
+
 Version 0.13 (release date: 2016-03-07)
 ---------------------------------------